refactor(ir): ✓ move async-static below the IR (#3113 S2)
async-static.ts holds three pure-syntactic predicates over `ts` nodes
(awaitIsStaticallyResolved, staticPromiseResolveSettledExpr,
unwrapPromiseTypeNode). Its own header says it was extracted as a LEAF
module "so the IR front-end (src/ir/select.ts, src/ir/from-ast.ts) can
consume them" — but the extraction left it in src/codegen/, so both of
those consumers imported UPWARD across the layer boundary.

It qualifies on the same criterion as slice 1's js-tag.ts: no codegen
imports, no CodegenContext, no codegen state. Its single import is
`src/ts-api.js`, a root-level module src/ir/ already depends on, so it is
dependency-free with respect to the ir/codegen boundary.

The three codegen consumers (index.ts, async-frame.ts, async-cps.ts —
which keeps its re-export for existing callers) now import down-stack,
the intended direction. This clears src/ir/select.ts of codegen imports
entirely: 20 files → 19.

Pure relocation, no logic change:
  - prove-emit-identity check → IDENTICAL, all 60 (file,target) emits
  - ts7 typecheck clean; async-await + regexp suites pass
  - ir-layering ratchet 94 → 92, banked in the baseline
